requirements.txt
# requirements.txt

# Server Dependencies
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
email-validator==2.0.0

# Client Dependencies
httpx==0.25.0
nest-asyncio==1.5.8
typing-extensions==4.8.0
asyncio==3.4.3

# Logging
python-json-logger==2.0.7

# LLM Model
# openai==1.3.0
llama-cpp-python==0.2.23
requests==2.31.0
ollama